Yordamchi/pyproject.toml
[project]
name = "python-template"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
requires-python = ">=3.11"
dependencies = [
    "ffmpeg==1.4",
    "pytelegrambotapi>=4.26.0",
    "python-dotenv>=1.0.1",
    "yt-dlp>=2025.1.26",
]
